--- a/src/csharp/Microsoft.Spark.MicroBenchmarks/Program.cs
+++ b/src/csharp/Microsoft.Spark.MicroBenchmarks/Program.cs
@@ -1,0 +1,18 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Diagnosers;
+using BenchmarkDotNet.Running;
+
+namespace Microsoft.Spark.MicroBenchmarks
+{
+    class Program
+    {
+        static void Main(string[] args) 
+            => BenchmarkSwitcher
+                .FromAssembly(typeof(Program).Assembly)
+                .Run(args, DefaultConfig.Instance.With(MemoryDiagnoser.Default));
+    }
+}