--- a/src/csharp/Microsoft.Spark.sln
+++ b/src/csharp/Microsoft.Spark.sln
@@ -1,4 +1,3 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28010.2046
@@ -23,6 +22,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Spark.CSharp.Examples", "..\..\examples\Microsoft.Spark.CSharp.Examples\Microsoft.Spark.CSharp.Examples.csproj", "{69F57F97-0198-4676-9BD9-E25407A132EA}"
 EndProject
 Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Microsoft.Spark.FSharp.Examples", "..\..\examples\Microsoft.Spark.FSharp.Examples\Microsoft.Spark.FSharp.Examples.fsproj", "{4E379DB3-7741-43C2-B32D-17AD96FEA7D0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Spark.MicroBenchmarks", "Microsoft.Spark.MicroBenchmarks\Microsoft.Spark.MicroBenchmarks.csproj", "{19F25D75-683A-4409-8EF0-A74CF3507D77}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -58,6 +59,10 @@ Global
 		{4E379DB3-7741-43C2-B32D-17AD96FEA7D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4E379DB3-7741-43C2-B32D-17AD96FEA7D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4E379DB3-7741-43C2-B32D-17AD96FEA7D0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{19F25D75-683A-4409-8EF0-A74CF3507D77}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{19F25D75-683A-4409-8EF0-A74CF3507D77}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{19F25D75-683A-4409-8EF0-A74CF3507D77}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{19F25D75-683A-4409-8EF0-A74CF3507D77}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE