--- a/src/csharp/Microsoft.Spark/Utils/PythonSerDe.cs
+++ b/src/csharp/Microsoft.Spark/Utils/PythonSerDe.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Diagnostics;
 using System.IO;
 using Microsoft.Spark.Sql;
@@ -23,6 +24,10 @@ namespace Microsoft.Spark.Utils
                 "pyspark.sql.types", "_parse_datatype_json_string", new StringConstructor());
             Unpickler.registerConstructor(
                 "pyspark.sql.types", "_create_row_inbound_converter", new RowConstructor());
+            BufferUnpickler.registerConstructor(
+                "pyspark.sql.types", "_parse_datatype_json_string", new StringConstructor());
+            BufferUnpickler.registerConstructor(
+                "pyspark.sql.types", "_create_row_inbound_converter", new RowConstructor());
         }
 
         /// <summary>
@@ -35,6 +40,14 @@ namespace Microsoft.Spark.Utils
             // Not making any assumptions about the implementation and hence not a class member.
             var unpickler = new Unpickler();
             var unpickledItems = unpickler.load(s);
+            Debug.Assert(unpickledItems != null);
+            return (unpickledItems as object[]);
+        }
+
+        internal static object[] GetUnpickledObjects(ReadOnlySpan<byte> memory)
+        {
+            var unpickler = new BufferUnpickler();
+            var unpickledItems = unpickler.load(memory);
             Debug.Assert(unpickledItems != null);
             return (unpickledItems as object[]);
         }