… src/csharp/Microsoft.Spark.MicroBenchmarks/Resources/serializedSampleInput.txt
… src/csharp/Microsoft.Spark/Interop/SparkEnvironment.cs