I have been analysing all the profiles and come up to the conclusion that deserialization, communication via sockets and serialization are the three most common perf bottlenecks in the functional benchmarks.

I did not want to rewrite the sockets to other IPC without further discussing it first so I started working on the deserialization perf.

After getting into the internals of existing unplickler I decided to rewrite it.

In this PR I am introducing a new type called `BufferUnpickler` which is circa a two times faster than `Unpickler`.

The idea is simple: we know the length of the message up front. Instead of using `MaxLengthReadStream` (which adds an overhead to every call to `ReadByte` and `Read`) and checking if we have enough of data in a few places I decided to read the bytes once, to an array from ArrayPool. So we do the IO once (instead of multiple calls to ReadByte and Read) and in most of the cases (when messageSize < ArrayPool.Shared.MaxSize) we don't allocate any managed memory. Morever, we check if we have enough of data only once, after reading it from the stream.

Once we have everything in a memory buffer we wrap in into `ReadOnlySpan<byte>` and pass to `BufferUnpickler`. `BufferUnpickler` processes the data in a similar way `Unpickler` did. The difference is that:
- `stream.ReadByte` is just `span[0]`
- `stream.Read(buffer, 0, size)` is just `span.Slice(0, size)`
This helps to avoid a lof of virtual method calls and `Array.Copy`.


The gain I got for deserializaiton of sample input from Q1 benchmark is x2.

|                                     Method |      Mean |     Error |    StdDev | Ratio |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------------------------------- |----------:|----------:|----------:|------:|-------:|-------:|------:|----------:|
| GetUnpickledObjectsFromMaxLengthReadStream | 10.921 us | 0.0163 us | 0.0145 us |  1.00 | 1.7242 | 0.1068 |     - |  10.65 KB |
|              GetUnpickledObjectsFromBuffer |  5.083 us | 0.0110 us | 0.0103 us |  0.47 | 1.7166 | 0.1068 |     - |   10.6 KB |

In all the profiles the deserialization takes also two times less time now:

![obraz](https://user-images.githubusercontent.com/6011991/56896792-0c408a00-6a8d-11e9-9d13-8c9e67017be2.png)

![obraz](https://user-images.githubusercontent.com/6011991/56896866-47db5400-6a8d-11e9-96f8-6d40fde8614b.png)

![obraz](https://user-images.githubusercontent.com/6011991/56897071-cf28c780-6a8d-11e9-8e27-568343bdd7fc.png)

![obraz](https://user-images.githubusercontent.com/6011991/56896984-97218480-6a8d-11e9-8176-9c27c312eb2a.png)