@adamsitnik 2019-04-29
I am going to close this PR and send a new one to https://github.com/irmen/Pyrolite

@eerhardt you are right, a lot of code duplication was unnecessary. I am going to reduce it without sacrificing the perf.